wip: refactor files
