+++ pages/about.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import Layout from "../components/Layout";
+import Layout from "../components/Layoutold";
 import "../components/Layoutdef.scss";
 import "../components/indexdef.scss";
